project/requirements/common.txt
click==6.7
Flask==0.12.2
gunicorn==19.7.1
icalendar==4.0.0
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
Werkzeug==0.12.2
todoist-python